test: Organize and renaming tests
